Add files via upload
一个快速的二次根式化简，输入数字<10^18,不使用素数打表

+++ quickly_simplify.java
@@ -9,7 +9,7 @@ public class quickly_simplify {
     int out = 1;
     int odd_max_root = (int)max_root>>1;
     odd_max_root = (odd_max_root<<1)+1;
-    int remain = (int)original;
+    long remain = original;
     if((original_sqrt-original_sqrt_int)==0) {
 	    return(String.valueOf(original_sqrt_int));
     }else {
